+++ server/api/project_admin_api.py
@@ -1,9 +1,10 @@
 from flask_restful import Resource, request, current_app
 from schematics.exceptions import DataError
+
 from server.models.dtos.project_dto import DraftProjectDTO, ProjectDTO
-from server.services.authentication_service import token_auth, tm
 from server.services.project_admin_service import ProjectAdminService, InvalidGeoJson, InvalidData, \
     ProjectAdminServiceError, NotFound
+from server.services.users.authentication_service import token_auth, tm
 from server.services.validator_service import ValidatorService
