+++ server/api/project_apis.py
@@ -1,9 +1,10 @@
 from flask_restful import Resource, current_app, request
 from schematics.exceptions import DataError
+
 from server.models.dtos.project_dto import ProjectSearchDTO
-from server.services.authentication_service import token_auth, tm
-from server.services.project_service import ProjectService, ProjectServiceError, NotFound
 from server.services.project_search_service import ProjectSearchService
+from server.services.project_service import ProjectService, ProjectServiceError, NotFound
+from server.services.users.authentication_service import token_auth, tm
 
 
 class ProjectAPI(Resource):
