+++ server/api/split_task_apis.py
@@ -1,9 +1,10 @@
 from flask_restful import Resource, current_app
-from server.models.dtos.grid_dto import SplitTaskDTO
 from schematics.exceptions import DataError
-from server.services.authentication_service import token_auth, tm
+
+from server.models.dtos.grid_dto import SplitTaskDTO
 from server.models.postgis.utils import NotFound
 from server.services.split_service import SplitService, SplitServiceError
+from server.services.users.authentication_service import token_auth, tm
 
 
 class SplitTaskAPI(Resource):
