+++ server/services/validator_service.py
@@ -8,7 +8,7 @@ from server.models.postgis.utils import NotFound, UserLicenseError
 from server.services.messaging.message_service import MessageService
 from server.services.project_service import ProjectService
 from server.services.stats_service import StatsService
-from server.services.user_service import UserService
+from server.services.users.user_service import UserService
 
 
 class ValidatatorServiceError(Exception):
