+++ tests/server/integration/services/test_authentication_service.py
@@ -1,10 +1,11 @@
 import os
 import unittest
 from urllib.parse import urlparse, parse_qs
-from tests.server.helpers.test_helpers import get_canned_osm_user_details
+
 from server import create_app
-from server.services.authentication_service import AuthenticationService
 from server.models.postgis.user import User
+from server.services.users.authentication_service import AuthenticationService
+from tests.server.helpers.test_helpers import get_canned_osm_user_details
 
 
 class TestAuthenticationService(unittest.TestCase):
